pyproject.toml
[project]
name = 'dify_plugin'
version = '0.8.0'
description = 'Dify Plugin SDK'
authors = [{ name = 'langgenius', email = 'hello@dify.ai' }]
dependencies = [
    'Flask~=3.1.3',
    'Werkzeug~=3.1.4',
    'dpkt~=1.9.8',
    'gevent~=25.5.1',
    'httpx~=0.28.1',
    'pydantic_settings>=2.5.0,<3.0.0',
    'pydantic>=2.8.2',
    'pyyaml~=6.0.1',
    'requests~=2.32.3',
    'socksio==1.0.0',
    'tiktoken~=0.8.0',
    'yarl>=1.9.4,<2.0',
    'packaging>=25.0',
    'urllib3>=2.6.3',  # common deps by plugins
    'aiohttp>=3.13.3',  # common deps by plugins
]
requires-python = '>=3.12'
readme = 'README.md'
license = 'Apache-2.0'
license-files = ['LICENSE']
keywords = ['dify', 'plugin', 'sdk']

[project.urls]
source = 'https://github.com/langgenius/dify-plugin-sdks'
homepage = 'https://github.com/langgenius/dify-plugin-sdks'
issues = 'https://github.com/langgenius/dify-plugin-sdks/issues'

[build-system]
requires = ['uv_build~=0.11']
build-backend = 'uv_build'

[dependency-groups]
dev = [
    'prek',
    'pytest',
    'pytest-cov',
    'pytest-mock',
    'pytest-xdist',
    'pytest-xprocess',
    'ruff',
    'ty',
]

[tool.uv]
default-groups = ['dev']

[tool.pytest]
minversion = '9.0'
addopts = ['-n', 'auto']
testpaths = ['tests']

[tool.ruff]
fix = true
preview = true
required-version = '>=0.15'

[tool.ruff.lint]
select = [
    'B',       # flake8-bugbear rules
    'C4',      # flake8-comprehensions
    'E',       # pycodestyle E rules
    'F',       # pyflakes rules
    'FURB',    # refurb rules
    'I',       # isort rules
    'N',       # pep8-naming
    'PT',      # flake8-pytest-style rules
    'PLC0208', # iteration-over-set
    'PLC0414', # useless-import-alias
    'PLR0402', # manual-from-import
    'PLR1711', # useless-return
    'PLR1714', # repeated-equality-comparison
    'RUF013',  # implicit-optional
    'RUF019',  # unnecessary-key-check
    'RUF100',  # unused-noqa
    'RUF101',  # redirected-noqa
    'RUF200',  # invalid-pyproject-toml
    'S506',    # unsafe-yaml-load
    'SIM',     # flake8-simplify rules
    'TRY400',  # error-instead-of-exception
    'TRY401',  # verbose-log-message
    'UP',      # pyupgrade rules
    'W191',    # tab-indentation
    'W605',    # invalid-escape-sequence

    # Security related linting rules
    # https://docs.astral.sh/ruff/rules/#flake8-bandit-s
    'S',

    # avoid shadowing builtins
    # https://docs.astral.sh/ruff/rules/#flake8-builtins-a
    'A',

    # prohibit calls to `breakpoint()` in committed code
    # https://docs.astral.sh/ruff/rules/#flake8-debugger-t10
    'T10',

    # disallow calling functions with boolean positional
    # arguments.
    #
    # https://docs.astral.sh/ruff/rules/boolean-positional-value-in-call/
    'FBT003',

    # Pycodestyle's warning
    # https://docs.astral.sh/ruff/rules/#warning-w
    'W',

    # Ruff specific rules
    # https://docs.astral.sh/ruff/rules/#ruff-specific-rules-ruf
    'RUF',
]

ignore = [
# ===================== Ruff-specific rules (RUF) =====================
    # ambiguous-unicode-character-string,
    # disabled beacause of false positives
    # ref: https://docs.astral.sh/ruff/rules/ambiguous-unicode-character-string/
    'RUF001',

    # ===================== flake8-bandit (S) rules =====================
    'S101', # use of assert detected, https://docs.astral.sh/ruff/rules/assert/
    # hardcoded-bind-all-interfaces, https://docs.astral.sh/ruff/rules/hardcoded-bind-all-interfaces/
    # disabled beacause of false positives
    'S104',
    # hardcoded-password-string, https://docs.astral.sh/ruff/rules/hardcoded-password-string/
    # disabled beacause of false positives
    'S105',
    'S404', # subprocess-import

    'B901', # return-in-generator
    'B903', # class-as-data-structure
    'E402', # module-import-not-at-top-of-file
    'N815', # mixed-case-variable-in-class-scope

    # Several package __init__ modules intentionally expose implementation
    # classes and runtime setup for the public SDK surface.
    'RUF067',

    # Disable UP007 for now due to the significant number of
    # violations in the current codebase.
    #
    # Reference: https://docs.astral.sh/ruff/rules/non-pep604-annotation-union/
    'UP007',
]

[tool.ruff.lint.pyflakes]
extend-generics = ['_pytest.monkeypatch', 'tests.integration_tests']
